Refactor de la clase fecha moviendola al paquete bot y casos de prueba
